fix: rename chart.ts into chart-type.ts to allow mobile app to work on macOS
