+++ packages/ui/src/components/templates/Dashboard/Chart/PieChart.tsx
@@ -22,7 +22,7 @@ import {PieChart as RNPieChart} from 'react-native-gifted-charts';
 import {useThemeColor} from '../../../../theme/ThemeContext';
 import {checkNullString} from '../../../../utils/strings';
 import {Text} from '../../../atoms';
-import Chart from '../chart';
+import Chart from '../chart-type';
 import {Data} from '../dashboard.helper';
 
 const MARGIN = 5;
